// src/@plugins/account/profile/profile/user/user.interface.ts
export interface UserInterface {
  _id?: string | Array<any>;
  username?: string | Array<any>;
  fullname: string | Array<any>;
  bio: string | Array<any>;
  email?: string | Array<any>;
  role?: string | Array<any>;
  password?: string | Array<any>;
  password_token?: string | Array<any>;
  permission_id?: boolean | string | Array<any>;
  _info?: any | Array<any>;
}